group service layer

+++ src/main/java/com/task/service/GroupService.java
@@ -1,29 +1,53 @@
 package com.task.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.task.converter.GroupConverter;
+import com.task.dto.GroupDtoForCreate;
 import com.task.entity.GroupEntity;
+import com.task.exception.GroupException;
 import com.task.repository.GroupRepository;
 
 @Service
 public class GroupService {
-
+	@Autowired
 	GroupRepository groupRepository;
 	
-	public GroupService(GroupRepository groupRepository) {
-		super();
-		this.groupRepository = groupRepository;
-	}
+	public GroupEntity addGroup(GroupDtoForCreate group) {
+		if (group != null) {
+			if (group.getName() != null) {
+				if (group.getDescription() != null) {
+						if(group.getCreatedAt() != null) {
+							GroupEntity groupToAdd = GroupConverter.toEntityForCreate(group);
+							groupRepository.addGroup(groupToAdd);
+						return groupToAdd;
+						} else {
+							throw new GroupException("Could not create group");
+						}
+				} else {
+						System.out.println("Group created at is mandatory");
+						throw new GroupException("Created at is required");
 
-	public void add(GroupEntity group) {
-		groupRepository.addGroup(group);
+					}
+			}else {
+						System.out.println("Group description is mandatory");
+						throw new GroupException("Description is required");
+
+					}
+		} else {
+				System.out.println("Group name is mandatory");
+				throw new GroupException("Name is required");
+		}
 	}
 	
-	public void update(GroupEntity group) {
-		groupRepository.updateGroup(group);
-	}
-	
-	public void delete(GroupEntity group) {
-		groupRepository.deleteGroup(group);
+	public void deleteGroup(long id) {
+		GroupEntity group = groupRepository.getGroupById(id);
+		if (group != null) {
+				groupRepository.deleteGroup(group);
+		} else {
+			System.out.println("Book not found");
+			throw new GroupException(id);
+		}
 	}
 }
